+++ src/Changes/Clutter.cs
@@ -10,6 +10,8 @@ public class Clutter : ManualChangesBase
   {
     var clutterTool = ItemManager["$PlumgaClutterTool"];
     var workbench = ZNetScene.instance?.GetPrefab("piece_workbench")?.GetComponent<CraftingStation>();
+    var forge = ZNetScene.instance?.GetPrefab("forge")?.GetComponent<CraftingStation>();
+    var stonecutter = ZNetScene.instance?.GetPrefab("piece_stonecutter")?.GetComponent<CraftingStation>();
     var hammer = ItemManager["Hammer"];
 
     // Remove some pieces
@@ -31,7 +33,7 @@ public class Clutter : ManualChangesBase
       piece.m_category = Piece.PieceCategory.Furniture;
       piece.m_comfort = 0;
       piece.m_craftingStation ??= workbench;
-      hammer.Pieces.Add(prefab);
+      if (!hammer.Pieces.Contains(prefab)) hammer.Pieces.Add(prefab);
     }
 
     // Remove Clutter Tool recipe
@@ -39,7 +41,7 @@ public class Clutter : ManualChangesBase
     ObjectDB.instance.m_recipes.Remove(clutterToolRecipe);
 
     // Adjust comfort
-    Dictionary<string, (int, Piece.ComfortGroup)> toAdjust = new()
+    Dictionary<string, (int, Piece.ComfortGroup)> comfortToAdjust = new()
     {
       ["$custompiece_coolchair"] = (1, Piece.ComfortGroup.Chair),
       ["$custompiece_couch1"] = (1, Piece.ComfortGroup.Chair),
@@ -57,10 +59,44 @@ public class Clutter : ManualChangesBase
       ["$custompiece_roundtablewithcloth"] = (1, Piece.ComfortGroup.Table),
 
     };
-    foreach (var prefab in hammer.Pieces.Where(p => toAdjust.ContainsKey(p.name)))
+    foreach (var prefab in hammer.Pieces.Where(p => comfortToAdjust.ContainsKey(p.name)))
     {
       var piece = prefab.GetComponent<Piece>();
-      (piece.m_comfort, piece.m_comfortGroup) = toAdjust[prefab.name];
+      (piece.m_comfort, piece.m_comfortGroup) = comfortToAdjust[prefab.name];
+    }
+
+    // Adjust required crafting table
+    Dictionary<string, CraftingStation> requiredCraftingTableToAdjust = new()
+    {
+      ["$custompiece_blackdragonstatuelarge"] = forge,
+      ["$custompiece_blackdragonstatuesmall"] = forge,
+      ["$custompiece_dragonstatuelarge"] = forge,
+      ["$custompiece_dragonstatuesmall"] = forge,
+      ["$custompiece_lokistatue"] = forge,
+      ["$custompiece_odinstatue"] = forge,
+      ["$custompiece_statuecorgi"] = stonecutter,
+      ["$custompiece_statuedeer"] = stonecutter,
+      ["$custompiece_statueevil_small"] = stonecutter,
+      ["$custompiece_statuehare"] = stonecutter,
+      ["$custompiece_statueseed"] = stonecutter,
+      ["$custompiece_celticidol1"] = stonecutter,
+      ["$custompiece_celticidol1small"] = stonecutter,
+      ["$custompiece_celticidol2"] = stonecutter,
+      ["$custompiece_celticidol3"] = stonecutter,
+      ["$custompiece_celticidol4"] = stonecutter,
+      ["$custompiece_celticidol5"] = stonecutter,
+      ["$custompiece_celticidol6"] = stonecutter,
+      ["$custompiece_celticidol7"] = stonecutter,
+      ["$custompiece_celticidol8"] = stonecutter,
+      ["$custompiece_celticidol9"] = stonecutter,
+      ["$custompiece_celticidol10"] = stonecutter,
+      ["$custompiece_celticidol11"] = stonecutter,
+      ["$custompiece_celticidol12"] = stonecutter,
+    };
+    foreach (var prefab in hammer.Pieces.Where(p => requiredCraftingTableToAdjust.ContainsKey(p.name)))
+    {
+      var piece = prefab.GetComponent<Piece>();
+      piece.m_craftingStation = requiredCraftingTableToAdjust[prefab.name];
     }
   }
 }