+++ src/Changes/FineWoodBuildPieces.cs
@@ -1,6 +1,7 @@
 using PieceManager;
 using LotusEcarlateChanges.Model.Reflection;
 using LotusEcarlateChanges.Model.Reflection.Plugins;
+using System.Collections.Generic;
 
 namespace LotusEcarlateChanges.Changes;
 
@@ -26,6 +27,26 @@ public class FineWoodBuildPieces : ReflectionChangesBase<FineWoodBuildPiecesPlug
       else if (piece.CustomCategory == "FineWoodFence") piece.Category = (int)BuildPieceCategory.Misc;
       // Relocate custom FineWoodPieces pieces to Building
       else if (piece.CustomCategory == "FineWoodPieces") piece.Category = (int)BuildPieceCategory.Building;
+    }
+
+    // Require Forge for heavy gate
+    plugin.PieceManager["BFP_HeavyGate"].Crafting.Set((int)CraftingTable.Forge);
+
+    // Require stonecutter for stone pieces
+    List<string> toAdjust = [
+      "BFP_StoneLightPost",
+      "BFP_StoneRoof26",
+      "BFP_StoneRoof45",
+      "BFP_StoneRoofICorner26",
+      "BFP_StoneRoofICorner45",
+      "BFP_StoneRoofOCorner26",
+      "BFP_StoneRoofOCorner45",
+      "BFP_StoneRoofTop26",
+      "BFP_StoneRoofTop45",
+    ];
+    foreach (var pieceName in toAdjust)
+    {
+      plugin.PieceManager[pieceName].Crafting.Set((int)CraftingTable.StoneCutter);
     }
   }
 }